feat(vif-p2p): Use link-local IPv6 as nexthop (#62)
Apparently, in IPv6 it's problematic to have a directly attached
prefix if the localhost has no IP in that prefix, as it is the
case when directly attaching a /128, i.e., a single IP. The
symptom we encountered is: When configuring bird6 to statically
route another prefix via that single IP, bird6 doesn't install the
route and instead complains that that IP (from the /128) is a
"strange next-hop".

Therefore, we instead of directly attaching the /128, we use the
link-local address of the VM as a hop. This requires to fix the
link-local address of the VM, as documented in `README.md`.